MF-1718 - Use static code analysis in CI (#1729)
* things, twins, and logger lint fixed


* all services updated, auth jwt not working, ineffectual assignment issue


* handle error from grpc server in endpointtest


* temp commit, auth/jwt needs to be resolved


* revert back to jwt v4 temporarily


* updated jwt tokenizer


* resolve EOF error for httptest requests


* fix auth jwt, update to registeredclaims


* fix ineffective assignment, auth/api/grpc endpoint failing


* temp commit, remove later


* fix grpc server setup


* resolve golangci tests, remove debug statements


* update golangci version and modify linters used


* fix failing tests


* fix grpc server for setup tests


* fix logging and errors inlined


* fix remarks, update grpc setup_test


* fix setup_test


* update setup_test grpc


* fix data race


* update setup_test grpc


* fix grpc setup down to single simple function


* fix linting issues


* resolve pr comments


* fix tests, handle returned errors, go mod tidy vendor


* fix errors from new linters


---------

+++ coap/client.go
@@ -88,7 +88,7 @@ func (c *client) Handle(msg *messaging.Message) error {
 		return errors.Wrap(ErrOption, err)
 	}
 	opts = append(opts, message.Option{ID: message.Observe, Value: []byte{byte(c.observe)}})
-	opts, n, err = opts.SetObserve(buff, uint32(c.observe))
+	opts, n, err = opts.SetObserve(buff, c.observe)
 	if err == message.ErrTooSmall {
 		buff = append(buff, make([]byte, n)...)
 		opts, _, err = opts.SetObserve(buff, uint32(c.observe))
